add all mentioned changes
