+++ pages/create.js
@@ -1,5 +1,27 @@
 import CardForm from "@/components/CardForm";
+import useSWR from "swr";
+import { useRouter } from "next/router";
 
 export default function Create() {
-  return <CardForm />;
+  const { mutate } = useSWR("/api/activities");
+  const router = useRouter();
+
+  async function handleSubmit(event) {
+    event.preventDefault();
+    const formData = new FormData(event.target);
+    const activityData = Object.fromEntries(formData);
+    const response = await fetch("/api/activities", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(activityData),
+    });
+    if (response.ok) {
+      mutate();
+      event.target.reset();
+      router.push("/");
+    }
+  }
+  return <CardForm onSubmit={handleSubmit} />;
 }
